refactor: move statement stub out of test class
